requirements.txt
pyyaml==6.0.2
python-frontmatter==1.1.0
tqdm==4.67.1
openai>=1.0.0
numpy>=1.26.0
python-dotenv>=1.0.1
pydantic>=2.7.0
streamlit>=1.35.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-chroma>=0.1.0
langchain-text-splitters>=0.3.0
langsmith>=0.1.0
fastapi
uvicorn
itsdangerous